bacon: turn on TARGET_USE_COMPAT_GRALLOC_ALIGN
For 8974 this boils down to whether to use get_gpu_pixel_alignment
from libadreno_utils.so, which is only used for the NV21 camera
preview format and is returning 32 instead of 16. This is
incorrect, but I'm sure this preexisting conditional compile
exists on 8974 for precisely this reason.

Fixes messed up 480p camera preview (CTS Verifier, barcode
scanner, 480p video recording, etc)